# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UrCampusFellowship API",
        "version": "1.0.0",
        "description": "Complete API documentation covering Student, Head, Admin, Public, and Upload routes."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        },
        {
            "url": "https://urcampus-fellowship.vercel.app",
            "description": "Staging Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sb-localhost-auth-token",
                "description": "Supabase Auth Cookie (handled automatically by Next.js)"
            }
        }
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Login with Email & Password",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "head@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123!"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
...
        "/api/student/register": {
            "post": {
                "summary": "Register a student to an active chapter",
                "tags": [
                    "Student"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "chapterId",
                                    "name",
                                    "phone",
                                    "program",
                                    "hall",
                                    "level",
                                    "campusId"
                                ],
                                "properties": {
                                    "chapterId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "program": {
                                        "type": "string"
                                    },
                                    "hall": {
                                        "type": "string"
                                    },
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "100",
                                            "200",
                                            "300",
                                            "400",
                                            "postgrad"
                                        ]
                                    },
                                    "campusId": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/student/waitlist": {
            "post": {
                "summary": "Join waitlist for a coming_soon chapter",
                "tags": [
                    "Student"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "chapterId",
                                    "email"
                                ],
                                "properties": {
                                    "chapterId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}